grsec: Apply grsecurity-2.9.1-2.6.32.60-201301021806.patch
commit b4d061468c5beb753d9d455e3ddb0d1f732865ab
Author: Brad Spengler <spender@grsecurity.net>
Date:   Tue Jan 1 14:10:27 2013 -0500

    Fix compiler warning

 lib/vsprintf.c |    2 +-
 1 files changed, 1 insertions(+), 1 deletions(-)

Signature-tree: 607d6dfbf5a0b9533295434c452ead48e84ba187

+++ lib/vsprintf.c
@@ -875,7 +875,7 @@ static char *pointer(const char *fmt, char *buf, char *end, void *ptr,
 	   as in the /proc/kallsyms case, as we make it display nothing
 	   for non-root users, and the real contents for root users
 	*/
-	if (ptr > TASK_SIZE && *fmt != 'P' && is_usercopy_object(buf)) {
+	if ((unsigned long)ptr > TASK_SIZE && *fmt != 'P' && is_usercopy_object(buf)) {
 		printk(KERN_ALERT "grsec: kernel infoleak detected!  Please report this log to spender@grsecurity.net.\n");
 		dump_stack();
 		ptr = NULL;
